Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
-int main()
+float main()
 {
     float a,r;
     scanf("%f",&a);
